requirements.txt
fastapi
uvicorn
python-telegram-bot==13.15
python-dotenv
urllib3==1.26.16